requirements.txt
flask
flask_sqlalchemy
flask_login
flask_wtf
python-dotenv
flask_migrate
psycopg2-binary
werkzeug